requirements.txt
torch
numpy
pillow
matplotlib
pathlib
requests